# api/swagger/swagger.yaml
swagger: "2.0"
info:
  version: "0.0.1"
  title: Hello World App
# during dev, should point to your local machine
host: localhost:10010
...
basePath: /
...
schemes:
  # TODO: remove http to make production-grade
  - http
  - https
# format of bodies a client can send (Content-Type)
consumes:
  - application/json; multipart/form-data
# format of the responses to the client (Accepts)
produces:
  - application/json; charset=utf-8; text/plain
paths:
...
  /issues/{_id}:
    x-swagger-router-controller: issues
    get:
      summary: Gets an issue by ID.
      operationId: getByID
      tags:
        - issues
      produces:
        - application/json
      parameters:
        - name: _id
          in: path
          description: Issue ID
          type: string
          # format: mongo objectid
          required: true
      responses:
        200:
          description: OK
          schema:
            $ref: '#/definitions/Issue'
...
definitions:
...
  Issue:
    type: object
    properties:
      _id:
        type: string
      name:
        type: string
    required:
      - _id
      - name
  # IssuesIndexResponse:
  #   # required:
  #   #   - _id
  #   properties:
  #     _id:
  #       type: string
...